updated stars and rain
