--- Collaboration2/Collaboration2/stars's animation/Stars.swift
@@ -1,8 +1,0 @@
-//
-//  Stars.swift
-//  Collaboration2
-//
-//  Created by vano Kvakhadze on 12.06.24.
-//
-
-import Foundation
